refactor(api): split universal upload + complete into vss_agents/api/videos.py
Split the generic chunk-proxy and completion routes out of
video_search_ingest.py into a new vss_agents/api/videos.py with two
explicit registration functions:

- register_video_upload          -> POST /api/v1/videos/chunked/upload
- register_video_upload_complete -> POST /api/v1/videos/{filename}/upload-complete

The naming makes the call boundary explicit: /upload-complete is called
for every profile, and on search profiles its hook drives ingestion
(RTVI-CV register + embedding generation). Each downstream step
self-skips when its backing service isn't configured, so base/alerts/
lvs profiles share the same endpoints with no per-profile YAML changes.

Backward compatibility:

- /api/v1/videos/{filename}/complete is registered as a deprecated
  alias for /upload-complete (same handler) so previously shipped UI
  builds keep working.
- /api/v1/videos-for-search/* PUT and /complete remain registered for
  the Video Management UI tab; FastAPI flags them deprecated=True in
  the OpenAPI schema. The post-processing helper they reuse now lives
  in videos.py and is imported back.

Tests:

- New test_videos.py: response/input model tests, _parse_optional_http_url,
  _run_post_upload_processing, the /upload-complete + /complete alias
  routes (handler-equivalence + deprecated flag), chunk-proxy route, and
  registration paths (skip-on-missing-VST, register-without-RTVI).
- test_video_search_ingest.py: trimmed to search-only coverage, asserts
  every videos-for-search route carries deprecated=True.
- test_custom_fastapi_worker.py: dispatcher tests now assert the two
  universal register fns fire on every profile (no capability flag),
  while videos-for-search remains gated.

UI: notifyGenericUploadComplete now POSTs to /upload-complete (the
canonical path); the deprecated /complete alias keeps older builds
working.
