+++ services/agent/src/vss_agents/api/video_search_ingest.py
@@ -14,28 +14,35 @@
 # limitations under the License.
 
 """
-Custom video ingest endpoint for VSS Search.
-This bypasses NAT's standard endpoint pattern to support direct file upload to VST.
+Deprecated: search-specific video ingest endpoints.
+
+These ``/api/v1/videos-for-search/*`` routes predate the universal video
+upload flow in ``vss_agents.api.videos`` and are retained only for backward
+compatibility with the Video Management UI tab, which still uploads chunks
+directly to VST and notifies the agent here. New callers should use the
+generic chunk-proxy + ``upload-complete`` pair from ``videos.py`` instead;
+both register on every profile and the ``upload-complete`` hook handles
+ingestion when RTVI is configured.
+
+Routes (all marked ``deprecated=True`` in the OpenAPI schema):
+    PUT  /api/v1/videos-for-search/{filename}            — streamed PUT upload
+    POST /api/v1/videos-for-search/{filename}/complete   — post-processing hook
 """
 
-import json
 import logging
 import os
 from typing import Any
-import urllib.parse
 
 from fastapi import APIRouter
 from fastapi import FastAPI
 from fastapi import HTTPException
 from fastapi import Request
 import httpx
-from pydantic import BaseModel
-from pydantic import Field
 
-from vss_agents.tools.vst.timeline import get_timeline
-from vss_agents.tools.vst.utils import VSTError
+from vss_agents.api.videos import VideoIngestResponse
+from vss_agents.api.videos import VideoUploadCompleteInput
+from vss_agents.api.videos import _run_post_upload_processing
 from vss_agents.utils.time_measure import TimeMeasure
-from vss_agents.utils.url_translation import rewrite_url_host
 
 logger = logging.getLogger(__name__)
 
@@ -46,241 +53,6 @@ ALLOWED_VIDEO_TYPES = {
 }
 
 
-def _parse_optional_http_url(url: str | None) -> urllib.parse.ParseResult | None:
-    """
-    Parse an optional HTTP(S) URL used to locate a downstream service.
-
-    Returns the parsed URL if it has a hostname, otherwise None. Catches
-    URLs like "", "http://", "http:", "http://host:" (no port body) —
-    anything that wouldn't successfully connect — and classifies them as
-    "not configured" so callers can skip the downstream step.
-
-    A URL relying on the scheme's default port (e.g. "http://host") is
-    considered valid: hostname alone is enough to connect.
-    """
-    if not url:
-        return None
-    try:
-        parsed = urllib.parse.urlparse(url)
-    except Exception:  # pragma: no cover — urlparse is extremely permissive
-        return None
-    if not parsed.hostname:
-        return None
-    # `http://host:` (trailing colon with empty port body) reaches us with a
-    # valid hostname but no resolvable port — Python's urlparse leaves the
-    # netloc as `host:` so calls would silently fall back to the scheme's
-    # default port (80 / 443) and connect to nothing. Treat that as
-    # misconfigured so callers skip the downstream step.
-    if parsed.netloc.endswith(":"):
-        return None
-    return parsed
-
-
-class VideoIngestResponse(BaseModel):
-    """Response for video ingest endpoint."""
-
-    message: str = Field(..., description="Status message indicating completion")
-    video_id: str = Field(..., description="The video ID used for storage")
-    filename: str = Field(..., description="The filename returned by VST after upload")
-    chunks_processed: int = Field(default=0, description="Number of chunks processed")
-
-
-class VideoUploadCompleteInput(BaseModel):
-    """Input for the upload-complete endpoint (chunked upload post-processing).
-
-    The UI forwards the full video-storage upload response as the request body
-    so it stays decoupled from the storage API shape. We extract the single
-    field the post-processing pipeline needs (sensorId) and ignore the rest.
-    Accepts both the camelCase ``sensorId`` (as VST returns it) and the
-    snake_case ``sensor_id`` for backward compatibility.
-    """
-
-    model_config = {"populate_by_name": True, "extra": "ignore"}
-
-    sensor_id: str = Field(
-        ...,
-        alias="sensorId",
-        min_length=1,
-        description="Video stream identifier from the upload response",
-    )
-
-
-async def _run_post_upload_processing(
-    camera_name: str,
-    sensor_id: str,
-    filename: str,
-    vst_url: str,
-    rtvi_embed_base_url: str,
-    rtvi_cv_base_url: str = "",
-    rtvi_embed_model: str = "cosmos-embed1-448p",
-    rtvi_embed_chunk_duration: int = 5,
-) -> VideoIngestResponse:
-    """
-    Run post-upload processing: get timeline, get video URL, add to RTVI-CV, generate embeddings.
-
-    Shared between the streaming PUT endpoint (small files, streamed through agent) and
-    the chunked-upload /complete endpoint (large files, uploaded directly to VST in chunks).
-
-    Args:
-        camera_name: Identifier sent as RTVI-CV ``camera_name``. Callers should pass
-            the filename without extension so the value is stable regardless of
-            which upload path was used. Note this is distinct from ``sensor_id``;
-            the returned ``VideoIngestResponse.video_id`` is set to ``sensor_id``,
-            not to ``camera_name``.
-        sensor_id: Stream id returned by VST after upload. Used for timeline
-            lookup, storage URL resolution, and as the ``video_id`` in the
-            response.
-        filename: Original filename (with extension). Used only in the human-
-            readable response message.
-    """
-    start_timestamp = "2025-01-01T00:00:00.000Z"
-
-    # Get timeline
-    try:
-        with TimeMeasure("video_ingest: get timeline from VST"):
-            timeline_start_time, timeline_end_time = await get_timeline(sensor_id, vst_url)
-    except VSTError as e:
-        logger.error("Timelines API failed for stream %s: %s", sensor_id, e)
-        raise HTTPException(status_code=502, detail=f"Timelines API failed: {e}") from e
-
-    if not timeline_start_time or not timeline_end_time:
-        error_msg = f"No valid timeline for stream {sensor_id}"
-        logger.error(error_msg)
-        raise HTTPException(status_code=502, detail=error_msg)
-
-    logger.info(
-        "Timeline for stream %s: start=%s, end=%s",
-        sensor_id,
-        timeline_start_time,
-        timeline_end_time,
-    )
-
-    # Get video URL via storage API
-    storage_url = f"{vst_url}/vst/api/v1/storage/file/{sensor_id}/url"
-    storage_params = {
-        "startTime": timeline_start_time,
-        "endTime": timeline_end_time,
-        "container": "mp4",
-        "configuration": json.dumps({"disableAudio": True}),
-    }
-    logger.info(f"Calling Storage API: GET {storage_url}")
-
-    async with httpx.AsyncClient(timeout=60.0) as client:
-        with TimeMeasure("video_ingest: get storage URL from VST"):
-            storage_response = await client.get(storage_url, params=storage_params)
-
-        if storage_response.status_code != 200:
-            error_msg = f"Storage API failed with status {storage_response.status_code}: {storage_response.text}"
-            logger.error(error_msg)
-            raise HTTPException(status_code=502, detail=f"Storage API failed: {error_msg}")
-
-        storage_result = storage_response.json()
-        vst_file_path = storage_result.get("videoUrl")
-        if not vst_file_path:
-            error_msg = f"Storage API response missing 'videoUrl' field: {storage_result}"
-            logger.error(error_msg)
-            raise HTTPException(status_code=502, detail=f"Storage API response invalid: {error_msg}")
-
-        logger.info(f"VST video URL obtained: {vst_file_path}")
-
-    # Add to RTVI-CV (if configured). The URL parser rejects empty, scheme-only,
-    # and "http://host:" (no port body) forms — anything that wouldn't connect.
-    parsed_cv = _parse_optional_http_url(rtvi_cv_base_url)
-    if parsed_cv is not None:
-        rtvi_cv_url = rtvi_cv_base_url.rstrip("/")
-        rtvi_cv_add_url = f"{rtvi_cv_url}/api/v1/stream/add"
-        rtvi_cv_payload = {
-            "key": "sensor",
-            "value": {
-                "camera_id": sensor_id,
-                "camera_name": camera_name,
-                "camera_url": vst_file_path,
-                "creation_time": start_timestamp,
-                "change": "camera_add",
-                "metadata": {"resolution": "1920x1080", "codec": "h264", "framerate": 30},
-            },
-            "headers": {"source": "vst", "created_at": start_timestamp},
-        }
-
-        logger.info(f"Adding video to RTVI-CV: POST {rtvi_cv_add_url}")
-
-        try:
-            async with httpx.AsyncClient(timeout=60.0) as rtvi_cv_client:
-                with TimeMeasure("video_ingest: register with RTVI-CV"):
-                    rtvi_cv_response = await rtvi_cv_client.post(rtvi_cv_add_url, json=rtvi_cv_payload)
-
-                if rtvi_cv_response.status_code not in (200, 201):
-                    error_msg = f"RTVI-CV returned {rtvi_cv_response.status_code}: {rtvi_cv_response.text}"
-                    logger.error(error_msg)
-                    raise HTTPException(status_code=502, detail=f"RTVI-CV add failed: {error_msg}")
-
-                logger.info(f"RTVI-CV video added: {sensor_id}")
-        except httpx.ConnectError:
-            logger.warning("RTVI-CV not reachable at %s, skipping (service may not be deployed)", rtvi_cv_add_url)
-        except httpx.TimeoutException:
-            logger.warning("RTVI-CV timed out at %s, skipping", rtvi_cv_add_url)
-    else:
-        logger.info("RTVI-CV not configured, skipping")
-
-    # Trigger embedding generation (skip if the embed service isn't configured).
-    # Uses the same parser as RTVI-CV for consistency — hostname-only URLs
-    # relying on the scheme's default port are accepted.
-    parsed_embed = _parse_optional_http_url(rtvi_embed_base_url)
-    chunks_processed = 0
-
-    if parsed_embed is None:
-        logger.info("RTVI Embed not configured, skipping embedding generation")
-    else:
-        rtvi_embed_url = rtvi_embed_base_url.rstrip("/")
-        embedding_url = f"{rtvi_embed_url}/v1/generate_video_embeddings"
-        parsed_vst = urllib.parse.urlparse(f"http://{vst_url}" if "://" not in vst_url else vst_url)
-        if not parsed_vst.hostname:
-            raise HTTPException(status_code=500, detail=f"Invalid vst_url format: {vst_url}")
-        translated_video_url = rewrite_url_host(vst_file_path, parsed_vst.hostname)
-        logger.info(f"Using internal VST URL for RTVI: {translated_video_url}")
-
-        embed_request = {
-            "url": translated_video_url,
-            "id": sensor_id,
-            "model": rtvi_embed_model,
-            "creation_time": start_timestamp,
-            "chunk_duration": rtvi_embed_chunk_duration,
-        }
-
-        logger.info(f"Calling RTVI Embedding API: POST {embedding_url}")
-
-        async with httpx.AsyncClient(timeout=600.0) as client:
-            with TimeMeasure("video_ingest: generate embeddings (RTVI)"):
-                embed_response = await client.post(
-                    embedding_url,
-                    json=embed_request,
-                    headers={"accept": "application/json", "Content-Type": "application/json"},
-                )
-
-            if embed_response.status_code != 200:
-                error_msg = (
-                    f"Embedding generation failed with status {embed_response.status_code}: {embed_response.text}"
-                )
-                logger.error(error_msg)
-                raise HTTPException(status_code=502, detail=f"Embedding generation failed: {error_msg}")
-
-            embed_result = embed_response.json()
-            logger.info("RTVI Embedding generation successful")
-            chunks_processed = embed_result.get("usage", {}).get("total_chunks_processed", 0)
-
-    message = (
-        f"Video {filename} successfully uploaded to VST and embeddings generated"
-        if parsed_embed is not None
-        else f"Video {filename} successfully uploaded to VST"
-    )
-    return VideoIngestResponse(
-        message=message,
-        video_id=sensor_id,
-        filename=filename,
-        chunks_processed=chunks_processed,
-    )
-
-
 def create_video_search_ingest_router(
     vst_internal_url: str,
     rtvi_embed_base_url: str,
@@ -289,29 +61,26 @@ def create_video_search_ingest_router(
     rtvi_embed_chunk_duration: int = 5,
 ) -> APIRouter:
     """
-    Create a FastAPI router for video search ingest.
+    Build the deprecated /api/v1/videos-for-search/* router.
 
-    This router handles raw binary data uploads and streams them directly
-    to VST without buffering the entire file in memory/disk.
-
-    Args:
-        vst_internal_url: Internal VST URL for API calls (required)
-        rtvi_embed_base_url: Base URL for RTVI Embed service (required)
-        rtvi_cv_base_url: Base URL for RTVI-CV service (optional, skipped if empty)
-        rtvi_embed_model: Model name for RTVI embedding generation (default: cosmos-embed1-448p)
-        rtvi_embed_chunk_duration: Chunk duration in seconds for embedding generation (default: 5)
-
-    Returns:
-        APIRouter with the video search ingest route
+    Routes are flagged ``deprecated=True`` in the OpenAPI schema. New code
+    should use the universal ``/api/v1/videos/chunked/upload`` +
+    ``/api/v1/videos/{filename}/upload-complete`` pair from
+    ``vss_agents.api.videos`` instead.
     """
     router = APIRouter()
 
     @router.put(
         "/api/v1/videos-for-search/{filename}",
         response_model=VideoIngestResponse,
-        summary="Upload video to VST",
-        description="Upload video file directly from client to VST without ANY intermediate storage",
+        summary="Upload video to VST (deprecated)",
+        description=(
+            "Deprecated: streamed PUT upload to VST. Prefer the universal "
+            "POST /api/v1/videos/chunked/upload + POST /api/v1/videos/"
+            "{filename}/upload-complete pair."
+        ),
         tags=["Video Ingest"],
+        deprecated=True,
     )
     async def upload_video_to_vst(
         filename: str,
@@ -330,18 +99,7 @@ def create_video_search_ingest_router(
         - Content-Type: allowed video MIME types (mp4, mkv)
         - Content-Length: <file_size>
         - Body: Raw binary video data
-
-        Args:
-            filename: Name of the video file (from URL path parameter)
-            request: FastAPI Request object for accessing raw stream
-
-        Returns:
-            VideoIngestResponse with upload status
-
-        Raises:
-            HTTPException: If upload fails
         """
-        # Fixed timestamp as per requirements
         start_timestamp = "2025-01-01T00:00:00.000Z"
 
         # Remove file extension if present to get video ID
@@ -387,8 +145,7 @@ def create_video_search_ingest_router(
             raise HTTPException(status_code=400, detail="Invalid Content-Length header") from e
 
         try:
-            # Stream directly from request to VST
-            # No intermediate storage, only 8KB in memory at a time
+            # Stream directly from request to VST (no intermediate storage).
             async with httpx.AsyncClient(timeout=300.0) as client:
                 logger.info(f"Streaming directly from client to VST at {vst_upload_url}")
 
@@ -399,19 +156,16 @@ def create_video_search_ingest_router(
                         headers={"Content-Type": content_type, "Content-Length": content_length},
                     )
 
-                # Check VST response
                 logger.info(f"VST upload response status: {vst_response.status_code}")
                 if vst_response.status_code not in (200, 201):
                     error_msg = f"VST upload failed with status {vst_response.status_code}: {vst_response.text}"
                     logger.error(error_msg)
                     raise HTTPException(status_code=502, detail=f"VST upload failed: {error_msg}")
 
-                # Parse VST response
                 vst_result = vst_response.json()
                 logger.info(f"VST upload successful - Streamed {content_length_int} bytes")
                 logger.debug(f"VST response body: {vst_result}")
 
-                # Extract streamId and sensorId from VST response
                 vst_sensor_id = vst_result.get("sensorId")
                 if not vst_sensor_id:
                     error_msg = f"VST response missing 'sensorId' field: {vst_result}"
@@ -420,11 +174,9 @@ def create_video_search_ingest_router(
 
                 logger.info(f"VST sensor ID: {vst_sensor_id}")
 
-                # Extract filename from VST response
                 vst_filename = vst_result.get("filename", filename)
                 logger.info(f"VST filename: {vst_filename}")
 
-                # Run post-upload processing (timeline, storage URL, RTVI-CV, embeddings)
                 return await _run_post_upload_processing(
                     camera_name=video_id,
                     sensor_id=vst_sensor_id,
@@ -445,14 +197,14 @@ def create_video_search_ingest_router(
     @router.post(
         "/api/v1/videos-for-search/{filename}/complete",
         response_model=VideoIngestResponse,
-        summary="Complete a chunked video upload and trigger post-processing",
+        summary="Complete a chunked video upload (deprecated)",
         description=(
-            "Called after a chunked upload directly to VST is finished. "
-            "Triggers timeline lookup, RTVI-CV registration, and embedding generation. "
-            "This bypasses the PUT upload endpoint to avoid Cloudflare's 100s timeout "
-            "for large files (the UI uploads chunks directly to VST, then calls this)."
+            "Deprecated: search-profile completion endpoint. Use POST "
+            "/api/v1/videos/{filename}/upload-complete instead — it works "
+            "across profiles and runs the same post-processing."
         ),
         tags=["Video Ingest"],
+        deprecated=True,
     )
     async def complete_video_upload(
         filename: str,
@@ -467,7 +219,7 @@ def create_video_search_ingest_router(
         """
         vst_url = vst_internal_url.rstrip("/")
         # Strip the file extension so RTVI-CV's camera_name matches the value
-        # the PUT upload path produces for the same filename (line ~322).
+        # the PUT upload path produces for the same filename.
         camera_name = filename.rsplit(".", 1)[0] if "." in filename else filename
 
         try:
@@ -490,22 +242,18 @@ def create_video_search_ingest_router(
     return router
 
 
-# This function will be called by custom FastAPI worker to register the router
 def register_video_search_ingest_routes(app: "FastAPI", config: "Any") -> None:
     """
-    Register videos-for-search routes to the FastAPI app.
+    Register the deprecated /api/v1/videos-for-search/* routes.
 
     Reads configuration from ``general.front_end.streaming_ingest`` in the YAML.
     The caller (``CustomFastApiFrontEndWorker``) gates this on the
-    ``enable_videos_for_search`` capability flag, so reaching this function
-    with a missing ``streaming_ingest`` is a programming error.
+    ``enable_videos_for_search`` capability flag.
 
-    These routes are search-only: both ``vst_internal_url`` and
-    ``rtvi_embed_base_url`` are required.
-
-    Args:
-        app: FastAPI application instance
-        config: NAT Config object containing application configuration
+    These routes are kept for backward compatibility with the Video Management
+    UI tab; new callers should use the generic
+    ``vss_agents.api.videos.register_video_upload`` +
+    ``register_video_upload_complete`` pair instead.
     """
     try:
         streaming_config = getattr(config.general.front_end, "streaming_ingest", None)
@@ -515,20 +263,19 @@ def register_video_search_ingest_routes(app: "FastAPI", config: "Any") -> None:
             )
 
         if streaming_config:
-            # streaming_ingest found in config (NAT supports extra fields)
             vst_internal_url = getattr(streaming_config, "vst_internal_url", None) or os.getenv("VST_INTERNAL_URL")
             rtvi_embed_base_url = getattr(streaming_config, "rtvi_embed_base_url", None)
             rtvi_cv_base_url = getattr(streaming_config, "rtvi_cv_base_url", None) or ""
             rtvi_embed_model = getattr(streaming_config, "rtvi_embed_model", "cosmos-embed1-448p")
             rtvi_embed_chunk_duration = getattr(streaming_config, "rtvi_embed_chunk_duration", 5)
-            logger.info("Using streaming_ingest config from YAML for search video ingest routes")
+            logger.info("Using streaming_ingest config from YAML for deprecated videos-for-search routes")
         else:
-            # Fallback: streaming_ingest not found (NAT strips unknown fields)
-            # Use environment variables. Require BOTH host AND port to build a
-            # URL — empty `RTVI_EMBED_PORT` (e.g. base profile, where RTVI isn't
-            # deployed) means RTVI is not configured and the URL stays empty,
-            # so /complete will register but skip the embedding step at request
-            # time instead of hanging on `http://host:` with no resolvable port.
+            # Fallback: streaming_ingest not in config (NAT may strip unknown
+            # sections). Use environment variables. Require BOTH host AND port
+            # to build a URL — empty `RTVI_EMBED_PORT` (e.g. base profile)
+            # means RTVI is not configured and the URL stays empty so the
+            # /complete handler will skip the embedding step at request time
+            # rather than hang on `http://host:`.
             vst_internal_url = os.getenv("VST_INTERNAL_URL")
             host_ip = os.getenv("HOST_IP")
             rtvi_embed_port = os.getenv("RTVI_EMBED_PORT", "")
@@ -548,10 +295,9 @@ def register_video_search_ingest_routes(app: "FastAPI", config: "Any") -> None:
             if streaming_config is not None:
                 logger.error("streaming_ingest configured but rtvi_embed_base_url is missing")
                 raise ValueError("rtvi_embed_base_url must be set when streaming_ingest is configured")
-            # Env-var fallback path with RTVI ports unset (e.g. base profile,
-            # where RTVI isn't deployed). Register the routes anyway so the UI
-            # gets 200 instead of 404; the /complete handler will skip the
-            # embedding step at request time when the URL is unset.
+            # Env-var fallback path with RTVI ports unset (e.g. base profile).
+            # Register the routes anyway so the UI gets 200 instead of 404;
+            # the /complete handler skips the embedding step at request time.
             logger.warning(
                 "RTVI not configured (HOST_IP=%r RTVI_EMBED_PORT=%r); "
                 "/complete routes will register but skip embedding generation",
@@ -567,186 +313,7 @@ def register_video_search_ingest_routes(app: "FastAPI", config: "Any") -> None:
             rtvi_embed_chunk_duration=rtvi_embed_chunk_duration,
         )
         app.include_router(router)
-        logger.info("Registered videos-for-search routes")
+        logger.info("Registered deprecated /api/v1/videos-for-search/* routes")
     except Exception as e:
         logger.error(f"Failed to register videos-for-search routes: {e}", exc_info=True)
-        raise
-
-
-# Forwarded to VST verbatim. Incoming request headers we don't want to pass
-# through (hop-by-hop, transport-level, or set by httpx itself).
-_CHUNK_PROXY_SKIP_HEADERS = {
-    "host",
-    "content-length",
-    "accept-encoding",
-    "connection",
-    "keep-alive",
-    "transfer-encoding",
-    "proxy-authenticate",
-    "proxy-authorization",
-    "te",
-    "trailers",
-    "upgrade",
-}
-
-
-def create_generic_video_router(
-    vst_internal_url: str,
-    rtvi_embed_base_url: str = "",
-    rtvi_cv_base_url: str = "",
-    rtvi_embed_model: str = "cosmos-embed1-448p",
-    rtvi_embed_chunk_duration: int = 5,
-) -> APIRouter:
-    """
-    Create a FastAPI router for the generic (profile-agnostic) video upload flow.
-
-    Used by the Chat upload path, which calls only the agent — not VST directly —
-    to stay backend-agnostic. The agent proxies each chunk to VST's nvstreamer
-    endpoint, then runs post-processing via the shared
-    `_run_post_upload_processing` helper.
-
-    Routes:
-        POST /api/v1/videos/chunked/upload    — proxies a single chunk to VST
-        POST /api/v1/videos/{filename}/complete — post-processing (timelines,
-             RTVI-CV register, embeddings). Each post-processing step skips
-             gracefully if its backing service isn't configured, so this route
-             works on base/alerts/lvs profiles too.
-    """
-    router = APIRouter()
-
-    @router.post(
-        "/api/v1/videos/chunked/upload",
-        summary="Proxy a chunked upload to VST (nvstreamer protocol)",
-        tags=["Video Ingest"],
-    )
-    async def proxy_chunk_to_vst(request: Request) -> dict[str, Any]:
-        """
-        Forward an incoming chunk (multipart body + nvstreamer-* headers) to VST's
-        /vst/api/v1/storage/file. Returns VST's response body and status to the
-        client unchanged.
-        """
-        vst_url = vst_internal_url.rstrip("/")
-        vst_chunk_url = f"{vst_url}/vst/api/v1/storage/file"
-
-        # Stream the request body straight through — buffering a 10MB chunk is
-        # fine, but read-all keeps memory bounded by the chunk size (not the
-        # file size).
-        body = await request.body()
-
-        headers = {k: v for k, v in request.headers.items() if k.lower() not in _CHUNK_PROXY_SKIP_HEADERS}
-
-        try:
-            async with httpx.AsyncClient(timeout=90.0) as client:
-                with TimeMeasure("video_ingest: proxy chunk to VST"):
-                    vst_response = await client.post(vst_chunk_url, content=body, headers=headers)
-        except httpx.ConnectError as e:
-            logger.error("VST not reachable at %s: %s", vst_chunk_url, e)
-            raise HTTPException(status_code=502, detail=f"VST not reachable: {e}") from e
-        except httpx.TimeoutException as e:
-            logger.error("VST timed out at %s: %s", vst_chunk_url, e)
-            raise HTTPException(status_code=504, detail=f"VST timed out: {e}") from e
-
-        # Forward VST's response shape verbatim (JSON for success, same status code).
-        try:
-            payload: dict[str, Any] = vst_response.json()
-        except ValueError:
-            # VST returned non-JSON — fall back to text so the client sees the error body.
-            raise HTTPException(
-                status_code=vst_response.status_code,
-                detail=f"VST returned non-JSON response: {vst_response.text[:500]}",
-            ) from None
-
-        if vst_response.status_code >= 400:
-            raise HTTPException(status_code=vst_response.status_code, detail=payload)
-        return payload
-
-    @router.post(
-        "/api/v1/videos/{filename}/complete",
-        response_model=VideoIngestResponse,
-        summary="Run post-upload processing for a chunked upload (generic)",
-        description=(
-            "Generic completion endpoint used by the Chat upload flow. Accepts "
-            "the full storage-API upload response as the request body and runs "
-            "timeline fetch → storage URL resolution → optional RTVI-CV register "
-            "→ optional embedding generation. Each post-processing step skips "
-            "gracefully if its backing service isn't configured, so this route "
-            "works across profiles."
-        ),
-        tags=["Video Ingest"],
-    )
-    async def generic_complete(
-        filename: str,
-        body: VideoUploadCompleteInput,
-    ) -> VideoIngestResponse:
-        sensor_id = body.sensor_id
-        # Strip the extension so RTVI-CV's camera_name matches what the search
-        # profile's /videos-for-search/*/complete and streaming PUT produce.
-        camera_name = filename.rsplit(".", 1)[0] if "." in filename else filename
-
-        try:
-            return await _run_post_upload_processing(
-                camera_name=camera_name,
-                sensor_id=sensor_id,
-                filename=filename,
-                vst_url=vst_internal_url,
-                rtvi_embed_base_url=rtvi_embed_base_url,
-                rtvi_cv_base_url=rtvi_cv_base_url,
-                rtvi_embed_model=rtvi_embed_model,
-                rtvi_embed_chunk_duration=rtvi_embed_chunk_duration,
-            )
-        except HTTPException:
-            raise
-        except Exception as exc:
-            logger.error("Generic /complete failed for %s: %s", filename, exc, exc_info=True)
-            raise HTTPException(status_code=500, detail=f"Post-processing failed: {exc}") from exc
-
-    return router
-
-
-def register_generic_video_routes(app: "FastAPI", config: "Any") -> None:
-    """
-    Register the generic (profile-agnostic) video upload routes.
-
-    Separate from `register_video_search_ingest_routes`: this one only needs
-    VST_INTERNAL_URL to register. Embedding and RTVI-CV URLs are passed
-    through when available and silently skipped at runtime when absent —
-    so base/alerts/lvs profiles get a working chunked-upload path too.
-    """
-    try:
-        streaming_config = getattr(config.general.front_end, "streaming_ingest", None)
-
-        if streaming_config:
-            vst_internal_url = getattr(streaming_config, "vst_internal_url", None) or os.getenv("VST_INTERNAL_URL")
-            rtvi_embed_base_url = getattr(streaming_config, "rtvi_embed_base_url", None) or ""
-            rtvi_cv_base_url = getattr(streaming_config, "rtvi_cv_base_url", None) or ""
-            rtvi_embed_model = getattr(streaming_config, "rtvi_embed_model", "cosmos-embed1-448p")
-            rtvi_embed_chunk_duration = getattr(streaming_config, "rtvi_embed_chunk_duration", 5)
-        else:
-            vst_internal_url = os.getenv("VST_INTERNAL_URL")
-            host_ip = os.getenv("HOST_IP")
-            rtvi_embed_port = os.getenv("RTVI_EMBED_PORT")
-            rtvi_cv_port = os.getenv("RTVI_CV_PORT")
-            rtvi_embed_base_url = f"http://{host_ip}:{rtvi_embed_port}" if host_ip and rtvi_embed_port else ""
-            rtvi_cv_base_url = f"http://{host_ip}:{rtvi_cv_port}" if host_ip and rtvi_cv_port else ""
-            rtvi_embed_model = os.getenv("RTVI_EMBED_MODEL", "cosmos-embed1-448p")
-            rtvi_embed_chunk_duration = 5
-
-        if not vst_internal_url:
-            logger.warning(
-                "VST_INTERNAL_URL not set — skipping generic video routes "
-                "(/api/v1/videos/chunked/upload and /api/v1/videos/{filename}/complete)",
-            )
-            return
-
-        router = create_generic_video_router(
-            vst_internal_url=vst_internal_url,
-            rtvi_embed_base_url=rtvi_embed_base_url,
-            rtvi_cv_base_url=rtvi_cv_base_url,
-            rtvi_embed_model=rtvi_embed_model,
-            rtvi_embed_chunk_duration=rtvi_embed_chunk_duration,
-        )
-        app.include_router(router)
-        logger.info("Successfully registered generic video routes (chunked upload proxy + /complete)")
-    except Exception as exc:
-        logger.error(f"Failed to register generic video routes: {exc}", exc_info=True)
         raise
